Artist not found error fix
During a search for an artist. if their account was either no longer active, deleted, or  the like. this would cause the program to throw an uncaught exception.

This has been fixed now and the program has been tested to work.

Added some extra logging for debug. Though end app is mostly untouched.